requirements.txt
autopep8==2.0.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
pycodestyle==2.9.1
requests==2.28.1
smmap==5.0.0
soupsieve==2.3.2.post1
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.10.0
